feat 비밀번호 채팅방 외부에서 입장 방지 * Modal창 UI/UX 이슈로 ChatList->Chat으로 기능 이관
