+++ frontend/src/Components/Chat/Chat.jsx
@@ -7,6 +7,8 @@ import ChatList from './ChatList'
 import GameList from './GameList'
 import Pagination from '../Pagenation/Pagination'
 import ChatRoomCreateModal from "./ChatRoomCreateModal";
+import ChatRoomPasswordModal from "./ChatRoomPasswordModal";
+import {useNavigate} from "react-router-dom";
 
 const Chat = () => {
     const currentUser = useContext(UserContext)
@@ -15,6 +17,9 @@ const Chat = () => {
     const [posts, setPosts] = useState(null)
     const [isLoading, setIsLoading] = useState(true)
     const [showCreateRoom, setShowCreateRoom] = useState(false)
+    const [showCheckPassword, setShowCheckPassword] = useState(false)
+    const [roomId, setRoomId] = useState(0)
+    const navigate = useNavigate()
 
     useEffect(() => {
         refreshList();
@@ -38,16 +43,36 @@ const Chat = () => {
         return <div>Loading...</div>
     }
 
-    const handleCreateRoom = async (roomId) => {
-        setShowCreateRoom(false);
+    const handleLinkClick = (e, post) => {
+        e.preventDefault();
 
-        await refreshList();
-        //TODO: room id에 해당하는 곳으로 navigate
+        if (post['is_secret']) {
+            // 비밀번호 인증 모달창 띄우기
+            setRoomId(post.id)
+            setShowCheckPassword(true)
+        } else {
+            // 채팅방으로 이동
+            const password = ''
+            navigate(`/chatroom/${post.id}`, {state: {password: password}})
+        }
+    };
+
+
+    const handleCreateRoom = async (roomId, password) => {
+        navigate(`/chatroom/${roomId}`, {state: {password: password}})
     }
 
-    const handleCloseRoom = async () => {
+    const handleCloseCreateRoom = async () => {
         setShowCreateRoom(false);
         await refreshList();
+    }
+
+    const handleEnterRoom = (roomId, password) => {
+        navigate(`/chatroom/${roomId}`, {state: {password: password}})
+    }
+
+    const handleClosePassword = () => {
+        setShowCheckPassword(false)
     }
 
     const handlePageChange = (page) => {
@@ -60,13 +85,16 @@ const Chat = () => {
 
     return (
         <div className={style.chat_container}>
+            {showCheckPassword && (
+                <ChatRoomPasswordModal onEnter={handleEnterRoom} onClose={handleClosePassword} roomId={roomId}/>
+            )}
             {showCreateRoom && (
-                <ChatRoomCreateModal onCreate={handleCreateRoom} onClose={handleCloseRoom}/>
+                <ChatRoomCreateModal onCreate={handleCreateRoom} onClose={handleCloseCreateRoom}/>
             )}
 
             <div className={style.chat_header}>
                 <button>새로고침</button>
-                <button onClick={()=>setShowCreateRoom(true)}>방만들기</button>
+                <button onClick={() => setShowCreateRoom(true)}>방만들기</button>
             </div>
 
             <div className={style.chat_category}>
@@ -84,7 +112,7 @@ const Chat = () => {
                 </button>
                 <hr/>
                 {view === 'chat' ? (
-                    <ChatList posts={posts} user={currentUser}/>
+                    <ChatList posts={posts} user={currentUser} onChatClick={handleLinkClick}/>
                 ) : (
                     <GameList posts={posts} user={currentUser}/>
                 )}
